+++ src/components/insights/mood/__tests__/mood-notes-timeline-removed.test.ts
@@ -4,6 +4,14 @@ import { join } from "node:path";
 
 const COMPONENT_DIR = join(__dirname, "..");
 const MOOD_LIST = join(__dirname, "../../../mood/mood-list.tsx");
+/**
+ * The two modules the mood insights surface is rendered from. The breakdown
+ * cluster lives behind `next/dynamic` in its own file; both are the surface.
+ */
+const INSIGHTS_SURFACE_FILES = [
+  join(COMPONENT_DIR, "mood-insights-sections.tsx"),
+  join(COMPONENT_DIR, "mood-insights-breakdowns.tsx"),
+];
 
 /**
  * v1.8.6 — the notes timeline display is removed from the mood insights
@@ -19,23 +27,31 @@ describe("mood notes timeline removal", () => {
   });
 
   it("no longer renders the notes timeline from the insights sections", () => {
-    const src = readFileSync(
-      join(COMPONENT_DIR, "mood-insights-sections.tsx"),
-      "utf8",
-    );
-    expect(src).not.toContain("MoodNotesTimeline");
-    expect(src).not.toContain("notesTimeline");
-    expect(src).not.toContain("notesTimelineTitle");
+    // Both halves of the surface: the eager module that paints the calendar
+    // and the assessment, and the deferred module that carries every
+    // breakdown. A guard reading only one of them would go quiet the moment
+    // the other grew a notes feed.
+    for (const file of INSIGHTS_SURFACE_FILES) {
+      const src = readFileSync(file, "utf8");
+      expect(src).not.toContain("MoodNotesTimeline");
+      expect(src).not.toContain("notesTimeline");
+      expect(src).not.toContain("notesTimelineTitle");
+    }
   });
 
   it("mounts the narrative feed in the insights sections", () => {
-    const src = readFileSync(
-      join(COMPONENT_DIR, "mood-insights-sections.tsx"),
-      "utf8",
-    );
     // v1.12.7 — the narrative one-liners now ride the merged "What stands out"
     // card (`MoodWhatStandsOut`), which renders `MoodNarrativeFeed` internally.
     // The structural guard tracks the mount point on the sections surface.
+    //
+    // v1.38 — that mount point moved into `mood-insights-breakdowns.tsx` when
+    // the below-the-fold cluster went behind `next/dynamic`. Read as the union
+    // of the two files, so the guard asserts the card is mounted SOMEWHERE on
+    // the surface rather than in one particular file — which is what it was
+    // always about.
+    const src = INSIGHTS_SURFACE_FILES.map((f) => readFileSync(f, "utf8")).join(
+      "\n",
+    );
     expect(src).toContain("MoodWhatStandsOut");
     expect(src).toContain("MoodInTargetTile");
   });
